refactor(header): improves code modularity on header scroll function
