+++ src/components/layout.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import React from 'react';
 import Footer from './footer';
 import Header from './header';
 
@@ -7,30 +7,19 @@ interface LayoutProps {
 }
 
 const Layout = ({ children }: LayoutProps) => {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(false);
+  // const [isSidebarOpen, setIsSidebarOpen] = useState(false);
 
-  function scrollToSection(event: React.SyntheticEvent) {
-    event.preventDefault();
-    const target = event.target as HTMLAnchorElement;
-    const id = target.getAttribute('href')?.replace('#', '');
-    const section = document.getElementById(String(id));
-
-    // section?.scrollIntoView({ behavior: "smooth", block: "center" })
-
-    if (section) {
-      const headerHeight = document.querySelector('header')?.clientHeight || 0;
-      const sectionPosition = section.offsetTop - headerHeight * 2;
-      window.scrollTo({ top: sectionPosition, behavior: 'smooth' });
-    }
-    setIsSidebarOpen(false);
-  }
+  // const handleScrollToSection = useCallback((event: React.SyntheticEvent) => {
+  //   scrollToSection(event);
+  //   setIsSidebarOpen(false);
+  // }, []);
 
   return (
     <div className="flex min-h-screen flex-col">
       <Header
-        scrollToSection={scrollToSection}
-        isSidebarOpen={isSidebarOpen}
-        setIsSidebarOpen={setIsSidebarOpen}
+      // scrollToSection={handleScrollToSection}
+      // isSidebarOpen={isSidebarOpen}
+      // setIsSidebarOpen={setIsSidebarOpen}
       />
       <div className="mx-8 my-6 flex-grow">{children}</div>
       <Footer />
